add useClickOutside hook
